@@ -5,19 +5,20 @@
     "build": "vite build",
     "preview": "npm run build && vite preview",
     "lint": "npm run check && npx eslint .",
-    "format": "npx @jill64/psx",
+    "format": "npx psvx",
     "check": "svelte-kit sync && npx svelte-check"
   },
   "prettier": "@jill64/prettier-config",
   "devDependencies": {
-    "@jill64/eslint-config-svelte": "1.3.7",
-    "@sentry/vite-plugin": "2.22.3",
-    "@sveltejs/adapter-cloudflare": "4.7.2",
-    "@sveltejs/adapter-vercel": "5.4.2",
-    "@sveltejs/adapter-netlify": "4.3.3",
-    "@sveltejs/adapter-auto": "3.2.4",
-    "svelte": "4.2.19",
-    "vite": "5.4.2",
-    "@sveltejs/vite-plugin-svelte": "3.1.2"
+    "@jill64/eslint-config-svelte": "2.0.2",
+    "@sentry/vite-plugin": "2.22.7",
+    "@sveltejs/adapter-cloudflare": "5.0.0",
+    "@sveltejs/adapter-vercel": "5.5.2",
+    "@sveltejs/adapter-netlify": "4.4.0",
+    "@sveltejs/adapter-auto": "3.3.1",
+    "@jill64/svelte-toast": "2.1.5",
+    "svelte": "5.16.0",
+    "vite": "6.0.6",
+    "@sveltejs/vite-plugin-svelte": "5.0.3"
   }
 }
