@@ -1,24 +1,25 @@
+import type { Span } from '@sentry/core'
 import {
   SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN,
+  SEMANTIC_ATTRIBUTE_SENTRY_SOURCE,
+  continueTrace,
   getActiveSpan,
-  getActiveTransaction,
   getCurrentScope,
-  getDynamicSamplingContextFromSpan,
-  runWithAsyncContext,
+  getDefaultIsolationScope,
+  getIsolationScope,
+  getTraceMetaTags,
   setHttpStatus,
-  spanToTraceHeader,
   startSpan,
-  captureException
+  winterCGRequestToRequestData,
+  withIsolationScope
 } from '@sentry/core'
-import type { Span } from '@sentry/types'
-import {
-  dynamicSamplingContextToSentryBaggageHeader,
-  objectify
-} from '@sentry/utils'
 import type { Handle, ResolveOptions } from '@sveltejs/kit'
 
-import { isHttpError, isRedirect } from '@sentry-sveltekit/common/utils.js'
-import { getTracePropagationData } from './utils.js'
+import {
+  flushIfServerless,
+  getTracePropagationData,
+  sendErrorToSentry
+} from './utils.js'
 
 export type SentryHandleOptions = {
   /**
@@ -55,34 +56,6 @@ export type SentryHandleOptions = {
   fetchProxyScriptNonce?: string
 }
 
-function sendErrorToSentry(e: unknown): unknown {
-  // In case we have a primitive, wrap it in the equivalent wrapper class (string -> String, etc.) so that we can
-  // store a seen flag on it.
-  const objectifiedErr = objectify(e)
-
-  // similarly to the `load` function, we don't want to capture 4xx errors or redirects
-  if (
-    isRedirect(objectifiedErr) ||
-    (isHttpError(objectifiedErr) &&
-      objectifiedErr.status < 500 &&
-      objectifiedErr.status >= 400)
-  ) {
-    return objectifiedErr
-  }
-
-  captureException(objectifiedErr, {
-    mechanism: {
-      type: 'sveltekit',
-      handled: false,
-      data: {
-        function: 'handle'
-      }
-    }
-  })
-
-  return objectifiedErr
-}
-
 /**
  * Exported only for testing
  */
@@ -110,28 +83,16 @@ export function addSentryCodeToPage(
   const nonce = fetchProxyScriptNonce ? `nonce="${fetchProxyScriptNonce}"` : ''
 
   return ({ html }) => {
-    // eslint-disable-next-line deprecation/deprecation
-    const transaction = getActiveTransaction()
-    if (transaction) {
-      const traceparentData = spanToTraceHeader(transaction)
-      const dynamicSamplingContext =
-        dynamicSamplingContextToSentryBaggageHeader(
-          getDynamicSamplingContextFromSpan(transaction)
-        )
-      const contentMeta = `<head>
-    <meta name="sentry-trace" content="${traceparentData}"/>
-    <meta name="baggage" content="${dynamicSamplingContext}"/>
-    `
-      const contentScript = shouldInjectScript
-        ? `<script ${nonce}>${FETCH_PROXY_SCRIPT}</script>`
-        : ''
+    const metaTags = getTraceMetaTags()
+    const headWithMetaTags = metaTags ? `<head>\n${metaTags}` : '<head>'
 
-      const content = `${contentMeta}\n${contentScript}`
+    const headWithFetchScript = shouldInjectScript
+      ? `\n<script ${nonce}>${FETCH_PROXY_SCRIPT}</script>`
+      : ''
 
-      return html.replace('<head>', content)
-    }
+    const modifiedHead = `${headWithMetaTags}${headWithFetchScript}`
 
-    return html
+    return html.replace('<head>', modifiedHead)
   }
 }
 
@@ -158,14 +119,36 @@ export function sentryHandle(handlerOptions?: SentryHandleOptions): Handle {
   }
 
   const sentryRequestHandler: Handle = (input) => {
-    // if there is an active transaction, we know that this handle call is nested and hence
-    // we don't create a new domain for it. If we created one, nested server calls would
-    // create new transactions instead of adding a child span to the currently active span.
-    if (getActiveSpan()) {
+    // event.isSubRequest was added in SvelteKit 1.21.0 and we can use it to check
+    // if we should create a new execution context or not.
+    // In case of a same-origin `fetch` call within a server`load` function,
+    // SvelteKit will actually just re-enter the `handle` function and set `isSubRequest`
+    // to `true` so that no additional network call is made.
+    // We want the `http.server` span of that nested call to be a child span of the
+    // currently active span instead of a new root span to correctly reflect this
+    // behavior.
+    // As a fallback for Kit < 1.21.0, we check if there is an active span only if there's none,
+    // we create a new execution context.
+    const isSubRequest =
+      typeof input.event.isSubRequest === 'boolean'
+        ? input.event.isSubRequest
+        : !!getActiveSpan()
+
+    if (isSubRequest) {
       return instrumentHandle(input, options)
     }
-    return runWithAsyncContext(() => {
-      return instrumentHandle(input, options)
+
+    return withIsolationScope((isolationScope) => {
+      // We only call continueTrace in the initial top level request to avoid
+      // creating a new root span for the sub request.
+      isolationScope.setSDKProcessingMetadata({
+        normalizedRequest: winterCGRequestToRequestData(
+          input.event.request.clone()
+        )
+      })
+      return continueTrace(getTracePropagationData(input.event), () =>
+        instrumentHandle(input, options)
+      )
     })
   }
 
@@ -180,31 +163,29 @@ async function instrumentHandle(
     return resolve(event)
   }
 
-  const { dynamicSamplingContext, traceparentData, propagationContext } =
-    getTracePropagationData(event)
-  getCurrentScope().setPropagationContext(propagationContext)
+  const routeName = `${event.request.method} ${
+    event.route?.id || event.url.pathname
+  }`
+
+  if (getIsolationScope() !== getDefaultIsolationScope()) {
+    getIsolationScope().setTransactionName(routeName)
+  }
 
   try {
     const resolveResult = await startSpan(
       {
         op: 'http.server',
         attributes: {
-          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.http.sveltekit'
+          [SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN]: 'auto.http.sveltekit',
+          [SEMANTIC_ATTRIBUTE_SENTRY_SOURCE]: event.route?.id ? 'route' : 'url',
+          'http.method': event.request.method
         },
-        name: `${event.request.method} ${
-          event.route?.id || event.url.pathname
-        }`,
-        status: 'ok',
-        ...traceparentData,
-        metadata: {
-          source: event.route?.id ? 'route' : 'url',
-          dynamicSamplingContext:
-            traceparentData && !dynamicSamplingContext
-              ? {}
-              : dynamicSamplingContext
-        }
+        name: routeName
       },
       async (span?: Span) => {
+        getCurrentScope().setSDKProcessingMetadata({
+          normalizedRequest: winterCGRequestToRequestData(event.request.clone())
+        })
         const res = await resolve(event, {
           transformPageChunk: addSentryCodeToPage(options)
         })
@@ -216,7 +197,9 @@ async function instrumentHandle(
     )
     return resolveResult
   } catch (e: unknown) {
-    sendErrorToSentry(e)
+    sendErrorToSentry(e, 'handle')
     throw e
+  } finally {
+    await flushIfServerless()
   }
 }
