@@ -1,4 +1,4 @@
-import type { NodeOptions } from '@sentry/node/types/index.js'
+import type { NodeOptions } from '../../sentry-javascript/packages/node/src/types.js'
 import type { DevOptions } from '../../types/DevOptions.js'
 import type { SentryHandleOptions } from '../server/handle.js'
 
